[proliferate] final frontend support debug cleanup
